templateing workerthread
