+++ include/feed/AprscFeed.h
@@ -23,9 +23,10 @@
 
 #include <cstdint>
 #include <memory>
+#include <string>
 
 #include "config/Properties.h"
-#include "util/WorkerThread.h"
+#include "util/WorkerThread.hpp"
 #include "util/defines.h"
 
 #include "Feed.h"
@@ -52,8 +53,8 @@ class AprscFeed : public Feed
 {
     static constexpr auto LOG_PREFIX = "(AprscFeed) ";
 
-    static parser::AprsParser s_parser;  ///< Parser to unpack response from Client
-    util::WorkerThread        m_worker;
+    static parser::AprsParser       s_parser;  ///< Parser to unpack response from Client
+    util::WorkerThread<std::string> m_worker;
 
 public:
     NOT_COPYABLE(AprscFeed)
